feat: ensure fonts initialise correctly
