+++ packages/widget-playground-next/src/app/page.tsx
@@ -14,7 +14,7 @@ import {
   WidgetConfigProvider,
   PlaygroundThemeProvider,
   DrawerControls,
-  FontLoadingProvider,
+  FontLoaderProvider,
 } from '@lifi/widget-playground';
 
 import { defaultWidgetConfig } from '@lifi/widget-playground/widget-config';
@@ -33,7 +33,7 @@ const AppProvider = ({ children }: PropsWithChildren) => {
         <EditToolsProvider>
           <WidgetConfigProvider defaultWidgetConfig={defaultWidgetConfig}>
             <PlaygroundThemeProvider>
-              <FontLoadingProvider>{children}</FontLoadingProvider>
+              <FontLoaderProvider>{children}</FontLoaderProvider>
             </PlaygroundThemeProvider>
           </WidgetConfigProvider>
         </EditToolsProvider>
