+++ packages/widget-playground/src/components/DrawerControls/DesignControls/FontsControl/fontDefinitions/googleFonts.ts
@@ -1,4 +1,4 @@
-import type { Font } from '../../../../../hooks';
+import type { Font } from '../../../../../providers';
 
 // NOTE: We could explore using the Google fonts developer API to write a
 //  script to auto generate more fonts we could even consider a light Next API endpoint to serve a list
