+++ packages/widget-playground/src/components/DrawerControls/DesignControls/FontsControl/fontDefinitions/systemFonts.ts
@@ -1,4 +1,4 @@
-import type { Font } from '../../../../../hooks';
+import type { Font } from '../../../../../providers';
 
 export const systemFonts: Font[] = [
   {
