+++ packages/widget-playground/src/hooks/index.ts
@@ -1,3 +1,1 @@
 export * from './useThemeMode';
-export * from './useFontLoader';
-export * from './types';
