+++ packages/widget-playground/src/index.ts
@@ -1,5 +1,4 @@
 export * from './providers';
 export * from './store';
-export * from './hooks';
 export * from './components/Widget';
 export * from './components/DrawerControls';
