+++ packages/widget-playground/src/providers/FontLoaderProvider/index.ts
@@ -1,0 +1,2 @@
+export * from './FontLoaderProvider';
+export * from './types';
