+++ packages/widget-playground/src/providers/index.ts
@@ -1,3 +1,4 @@
 export * from './PlaygroundThemeProvider';
 export * from './ExternalWalletProvider';
 export * from './EnvVariablesProvider';
+export * from './FontLoaderProvider';
